imx: check for this file first

+++ tools/install_image.sh
@@ -177,7 +177,9 @@ mmc_write_imx_bootlets () {
 	echo "Installing ${KERNEL_UTS}.sd_mmc_bootstream.raw to boot partition"
 	echo "-----------------------------"
 
-	sudo dd if="${DIR}/deploy/${KERNEL_UTS}.sd_mmc_bootstream.raw" of=${MMC}${PARTITION_PREFIX}${BOOT_PARITION}
+	if [ -f "${DIR}/deploy/${KERNEL_UTS}.sd_mmc_bootstream.raw" ] ; then
+		sudo dd if="${DIR}/deploy/${KERNEL_UTS}.sd_mmc_bootstream.raw" of=${MMC}${PARTITION_PREFIX}${BOOT_PARITION}
+	fi
 	sync
 	sync
 	mmc_find_rootfs
